Use contract specific ABI encoding
I replaced all occurrences of `encodeFunctionCall` with the correct calls
to `methods.myMethod().encodeABI()`.

The existing integration tests assert that the behavior is as desired. I
did not see the value in creating unit tests that would basically mirror
the methods' code and "assert" that the same things are the same.

Fixes #152